user date input stored and passed in AllEvents.js API call function
